+++ app/Controllers/ProductCollectionController.php
@@ -2,10 +2,10 @@
 namespace AgreableProductPlugin\Controllers;
 
 use Herbert\Framework\Models\Post;
-use \Timber;
-use \TimberPost;
-use \Exception;
-use \stdClass;
+use Timber;
+use TimberPost;
+use Exception;
+use stdClass;
 
 class ProductCollectionController {
   /**
@@ -19,25 +19,9 @@ class ProductCollectionController {
 
     $context = Timber::get_context();
     $context['product_collection'] = $product_collection;
-    $context['js_string'] = $this->get_javascript_string();
 
     Timber::render('@AgreableProductPlugin/intro.twig', $context, false);
   }
-
-  protected function get_javascript_string() {
-    $plugin_root = realpath(__DIR__ . '/../..');
-    $port_file = 'webpack-current-port.tmp';
-    $port_file_location = $plugin_root . '/' . $port_file;
-
-    if (getenv('WP_ENV') === 'development' && file_exists($port_file_location)) {
-      $port_number = file_get_contents($port_file_location);
-      return "<script src='http://localhost:$port_number/static/app.js'></script>";
-    }
-
-    return '<script>' . file_get_contents($plugin_root . '/resources/assets/app.js') .
-      '</script>';
-  }
-
 
   protected function get_product_collection_by_slug($product_collection_slug) {
     $args = array(
