+++ app/Controllers/ProductController.php
@@ -2,10 +2,10 @@
 namespace AgreableProductPlugin\Controllers;
 
 use Herbert\Framework\Models\Post;
-use \Timber;
-use \TimberPost;
-use \Exception;
-use \stdClass;
+use Timber;
+use TimberPost;
+use Exception;
+use stdClass;
 
 class ProductController {
    /**
@@ -19,24 +19,9 @@ class ProductController {
 
     $context = Timber::get_context();
     $context['product'] = $post;
-    $context['js_string'] = $this->get_javascript_string();
 
     Timber::render('@AgreableProductPlugin/single-product.twig', $context, false);
 
-  }
-
-  protected function get_javascript_string() {
-    $plugin_root = realpath(__DIR__ . '/../..');
-    $port_file = 'webpack-current-port.tmp';
-    $port_file_location = $plugin_root . '/' . $port_file;
-
-    if (getenv('WP_ENV') === 'development' && file_exists($port_file_location)) {
-      $port_number = file_get_contents($port_file_location);
-      return "<script src='http://localhost:$port_number/static/app.js'></script>";
-    }
-
-    return '<script>' . file_get_contents($plugin_root . '/resources/assets/app.js') .
-      '</script>';
   }
 
   protected function get_product_by_slug($product_slug) {
