+++ resources/views/base.twig
@@ -13,7 +13,7 @@
     {# {% if env == "development" %} #}
 
       <hr style="border: solid 1px red; ">
-      {{ js_string | raw }}
+      {{ product_plugin.js_string | raw }}
 
     {#
     {% else %}
